classifier/pyproject.toml
[project]
name = "classifier"
version = "0.1.0"
description = "Zero-shot multi-hypothesis classifier prototype"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "transformers>=4.44",
    "torch>=2.2",
    "python-dotenv>=1.0",
    "unidecode>=1.3",
]
